requirements.txt
fastapi
uvicorn
supabase
python-dotenv
jinja2
pubnub
